refactor: move some utils func to utils.py
